Layout_TestCase/http/tests/xmlhttprequest/access-control-basic-non-simple-deny-cached-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/access-control-basic-non-get-allow.cgi. Request header field Content-Type is not allowed by Access-Control-Allow-Headers.
PASS: Exception thrown. Cross-domain access is not allowed in 'send'. [NETWORK_ERR: XMLHttpRequest Exception 101].
